Fix skipping of first slot.
